backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pdfplumber
python-multipart